+++ tests/com/yanncebron/m68kplugin/lexer/MinimalLexerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 The Authors
+ * Copyright 2022 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -280,5 +280,18 @@ public class MinimalLexerTest extends M68kLexerTestCase {
         "id ('*')\n" +
         "- ('-')\n" +
         "dec_number ('42')");
+  }
+
+  public void testCurrentPcSymbolAfterOpeningParentheses() {
+    doTest(" dc.b (pc-2)",
+      "WHITE_SPACE (' ')\n" +
+        "dc ('dc')\n" +
+        ".b ('.b')\n" +
+        "WHITE_SPACE (' ')\n" +
+        "( ('(')\n" +
+        "pc ('pc')\n" +
+        "- ('-')\n" +
+        "dec_number ('2')\n" +
+        ") (')')");
   }
 }